comparison of different types

+++ YACReaderLibrary/yacreader_local_server.cpp
@@ -104,7 +104,7 @@ void YACReaderClientConnectionWorker::run()
 	int dataAvailable = 0;
 	QByteArray packageSize;
     clientConnection->waitForReadyRead(1000);
-	while(packageSize.size() < sizeof(quint32) && tries < 20)
+	while((unsigned int)packageSize.size() < sizeof(quint32) && tries < 20)
 	{
 		packageSize.append(clientConnection->read(sizeof(quint32) - packageSize.size()));
 		clientConnection->waitForReadyRead(100);
